requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
segmentation-models-pytorch
